rk3288: Some python for generating tables in dw_hdmi-rockchip.c
This script was written based on looking at the tables provided for
the PHY configuration and figuring out how to do it automatically.
This allows us to make fuller tables with less guesswork.

Technically this could go into the kernel itself, but for now we
probably just want to fill in the tables that the main dw_hdmi
expects.

This script currently only works for NO pixel repetition.

BUG=chrome-os-partner:41190
TEST=Looking at 8-bit tables, only 148.5MHz settings change.  They
change in a way that matches the docs and that makes sense.
TEST=Spot check many entries in the table and see that they match.

Change-Id: I235b34ebf59a680cf888cfe47807cf2315db3c94
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/285855
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>