bsdrm: add debug output for errors
Because of the high level nature of this library, it can be difficult to
pinpoint where or why something failed. This change adds some simple
debug output functions. It employs those functions only in places where
the return value of the function does not indicate what sort of error
happened. Think functions that have no return value or functions that
are doing multiple things that could fail but only return a single error
value.

The purpose of bsdrm is debugging issues, so there is no way to disable
or redirect the output, but it would be possible to change that if
desired in a later CL.

TEST=run stripe while ui is started, observe handy output on stderr
BUG=None

Change-Id: I523f69300457db3eeb94ad12ced536a756e97774
Reviewed-on: https://chromium-review.googlesource.com/326460
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>