bsdrm: make drm_fb module not do buffer allocation
It would be inconsistent to have drm_fb allocate buffers for the user
just because it will also use that buffer as a drm fb. The upcoming egl
module also needs to import buffers. It would be more consistent to have
the application make the buffer and ask drm_fb and egl to import that
given buffer. As it turns out, gbm_bo_create takes the same parameters
as bs_drm_fb_new, so users that want to allocate should use that
instead.

TEST=run stripe on a device
BUG=None

Change-Id: Ic2efc26178edb41b8370ca21822f95901f21dcee
Reviewed-on: https://chromium-review.googlesource.com/327332
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Haixia Shi <hshi@chromium.org>