bsdrm: add app class for ease of getting pictures on screen
The app class is meant to support common usage patterns of DRM apps.
With just a few calls and no configuration, it sets up all the
framebuffers and a good default mode and then displays them on screen.
The actual application can pull the framebuffers out of the app to do
rendering using EGL or mapping to user space for CPU rendering.

The app class is highly coupled to other classes within bsdrm because it
is meant to integrate many features of the library for ease of use. It
may not be possible to do your drm-test using bs_app, in which case you
should probably just skip using it.

BUG=None
TEST=drm-tests still builds

Change-Id: I5d3a9ca95da8b1018bb0cb0d1fc8f064f71b8d9f
Reviewed-on: https://chromium-review.googlesource.com/327674
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Haixia Shi <hshi@chromium.org>