bsdrm: add gl module of OpenGL helpers
The initial version of this module only has shader and program creation
helpers. This is something every OpenGL ES 2.0+ program needs to do and
it's both common and error prone.

An example of how to use the new module can be found in the updated
null_platform_test.c included in thsi CL.

BUG=chromium:587632
TEST=null_platform_test

Change-Id: I068512e2317034f004cdea1e766015658803593d
Reviewed-on: https://chromium-review.googlesource.com/329739
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Haixia Shi <hshi@chromium.org>