drm-tests: add linear_bo_test for testing GBM_BO_USE_LINEAR
The GBM_BO_USE_LINEAR usage flag in minigbm needs a test and this CL
provides it. This test uses buffers with several combinations of format
and usage and compares the success of creating those buffers with what
it expects. It will also display all the buffers that are successfully
made with a pattern that identifies the test case. This is to ensure the
tester can visually see that the pattern was rendered and displayed
properly. Flips, rotations, endian mismatch, and tiling mismatch would
all be obvious in the pattern used in this test.

TEST=linear_bo_test
BUG=None

Change-Id: I3692ccb900d0a398af25350a8ad5801f9c3ca2aa
Reviewed-on: https://chromium-review.googlesource.com/329932
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>