bsdrm: add pipe plumber class for making pipes with constraints
The bs_drm_pipe_plumber class is like an expanded version of
bs_drm_pipe_make. By default, it will check every card and open it on
behalf of the user. It also supports more constraints, and makes adding
more with later version of this library very easy without breaking
existing code. Currently, it supports limiting which connectors, crtcs,
and drm cards are used for making a pipe. The connectors can even be
ranked to support fall back connectors. Another advantage is that the
pipe plumber can be used to collect the drmModeConnector that was
already gotten during the pipe making process. It's purely an
optimization, but saves 1 second per drmModeGetConnector call in the
case that an external connector is used.

This change also includes a related module used for ranking connectors.
It includes some standard rankings that most apps should use.

The stripe example was changed to use this new class to show how it
works.

The main motivation for these changes was the gamma_test which will need
both the crtc masking and connector ranking. It was also used to
integrate the drm_open and drm_pipe functions which were starting to
have serious overlap. It no longer makes sense to keep them separate,
but this CL won't remove the drm_open module for now.

BUG=None
TEST=stripe

Change-Id: Iaad07803a02610a2142e958b5dfb6199f41b6b33
Reviewed-on: https://chromium-review.googlesource.com/336320
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Haixia Shi <hshi@chromium.org>