drm-tests: refactor gamma_test to use bsdrm
The gamma_test in particular exercised the advanced features of bsdrm
because it supports external connectors and CRTC masks. Using bsdrm has
made the test focus much more on what matters to it, generating and
testing gamma tables on CRTCs.

This test also used the "sp" library whose functionality can be
fulfilled better by gbm for making buffers and bsdrm for the mmaping and
display. This CL removes the dependency on "sp" for gamma_test.

TEST=gamma_test
BUG=chromium:587632

Change-Id: I8f0f88b45e27d1af88f160f53c8d258602106cc8
Reviewed-on: https://chromium-review.googlesource.com/336312
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Dominik Behr <dbehr@google.com>
Reviewed-by: Haixia Shi <hshi@chromium.org>