bsdrm: make framebuffers from multi-planar gbm_bo
Before this change, bs_drm_fb_create_fd_gbm would assume a single plane
when creating a framebuffer. With this change, up to four planes with
handles, offsets, and strides will be properly put into the framebuffer
object.

Note that this also fixes a small bug in which memory for only a single
handle was passed to drmModeAddFB2, instead of an array of 4.

BUG=None
TEST=stripe still works, nv12_test after the refactor lands

Change-Id: I0143d39bd287f183c798e2fddd5b0c4a1e37e3e1
Reviewed-on: https://chromium-review.googlesource.com/338963
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>