drm-tests: mv nv12_test to plane_test and support more formats
What's better than nv12_test? A test that can write to an unbounded
number of formats is what. This patch lays out the groundwork for easily
adding more formats to draw to. It also allows the user of plane_test to
pass in the desired format as the first argument.

TEST=plane_test
BUG=chromium:607241
CQ-DEPEND=CL:341262

Change-Id: Icdbed59258a85ec1b1eb0753e94b7f86e5d9669f
Reviewed-on: https://chromium-review.googlesource.com/341324
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>