drm-tests: gamma_test: don't require an active connector for every crtc
on Cyan, the gamma_test succeeds for the first two CRTCs but fails for the
last. That's because the only connectors that will work for CRTC2 have a
state of DRM_MODE_DISCONNECTED and have no modes associated with them.

This is a valid result from probing KMS, so we shouldn't fail the test
when we are unable to make a pipe for a particular crtc.

BUG=chromium:687351
TEST=gamma_test on cyan

Change-Id: I0d5405e9681247bc9caec25b5ee3cb9ca6276924
Reviewed-on: https://chromium-review.googlesource.com/446893
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>