drm-tests: test "rotation" KMS property.
KMS exposes a "rotation" property that can be use to rotate or to mirror
a HW overlays.

https://chromium-review.googlesource.com/c/468191/ introduces the property
on rockchip and adds support for the value of "reflect-y."

This CL adds a test to explicitly test "rotation" set to "reflect-y" on
the primary plane and on an additional overlay.

BUG=b:36752417
TEST="atomictest -t orientation -c 1" shows upside down patterns.

Change-Id: Ifa7b619b584e248a4c04ca10feeba5558711cf1a
Reviewed-on: https://chromium-review.googlesource.com/470606
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Dongseong Hwang <dongseong.hwang@intel.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>