drm-tests: draw.c: rename pixel skip to byte skip
Now we've changed how draw YUV surfaces, pixel_skip needs to mean
the number of bytes between the same draw component on the same
plane. For example for NV12, the number bytes between U components
on the second plane is 2. This is analogous to chroma step in
struct android_ycbcr.

BUG=b:37526302
TEST=plane_test (defaults to NV12) on minnie draws striped pattern

Change-Id: Iaff77bbbd5538d9c821e2cc52dc36d46249cbff8
Reviewed-on: https://chromium-review.googlesource.com/489811
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Dongseong Hwang <dongseong.hwang@intel.com>
Reviewed-by: Zach Reizner <zachr@chromium.org>