drm-test: plane_test: update default mapper and memset
This patch updates the default mapper to bs_mapper_gem_new()
from bs_mapper_dma_buf_new() so as to align with other
drm-tests like atomictest etc.

Also the patch updates the memsetting of the mapped gbm_bo
to black color as black is generally the color_key for blending,
by memsetting it to white the data on underlay can't be visible,
if only underlay plane is set.

BUG=b:37526302
TEST=plane_test --format AR24 --size 500x50 -p --format NV12 --size 1366x768
TEST=plane_test --format NV12 --size 500x500
TEST=plane_test --format AR24 --size 500x500
draws striped pattern

Change-Id: If812559b7d0e66998f128c059eab44445c8ef8bf
Signed-off-by: Shirish S <shirish.s@amd.com>
Reviewed-on: https://chromium-review.googlesource.com/527983
Reviewed-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>