drm-tests: vgem_test: Exercise import_sg_table
The import_foreign part of vgem test is meant to exercise the import and
export of buffers between vgem and other cards. The test however always
opens /dev/dri/card0, which in many cases is vgem itself. There is
short-circuit logic in drm_gem_prime_import which notices if the buffer
is being imported/exported to the same device, and exits early if so.

This change opens a device that has resources vgem doesn't have, to
ensure that a non-vgem card is used to exercise the full import code
path in the kernel. With this change, the kernel bug in 3.14
chromium:804428 can be observed.

BUG=chromium:804428
TEST=Added and observed printk in vgem_gem_prime_import_sg_table and
drm_gem_prime_import.
CQ-DEPEND=CL:879396

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: Id014cdbae34802783943d4675dd1ff71d7359382
Reviewed-on: https://chromium-review.googlesource.com/879398
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Evan Green <evgreen@google.com>
Reviewed-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>