drm-tests: set only supported formats for overlay planes
This patch removes hardocding of a fixed format for overlay
and primary planes.
With this patch a format that is supported by the plane is
identified and then tried to set hence ensuring the real
purpose of test i.e, atomicity is met, appropriately.

This patch also introduces a wrapper to initialise planes
for both yuv and rgb formats.

BUG=b:65297611
TEST=On Kahlee verified that the below tests passes:
	atomictest -t disable_primary
	(it fails without this patch)
	atomictest -t orientation
	atomictest -t plane_ctm

Change-Id: I816c121ebfd7ce476b453e9f301f3822283a650a
Signed-off-by: Shirish S <shirish.s@amd.com>
Reviewed-on: https://chromium-review.googlesource.com/885703
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>