drm-tests: Speed up the tests
Previously, test cases that moved across the screen had a usleep of
1e6 / 60, while tests that displayed a few frames (i.e, test_orientation)
had  sleep of 1e6.  The change to test_and_commit() made every case
have a usleep of 1e6. This changes the timeouts back, and adds an 'automatic'
option that runs with a usleep of 0.

BUG=chromium:828163
TEST=builds

Change-Id: Ife1889387fca9632cc37c9272cfc6115dbd1cdbe
Reviewed-on: https://chromium-review.googlesource.com/1020405
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>