drm-tests: Add test for CRTC GAMMA_LUT property
This adds a test for the drm property 'GAMMA_LUT' that allows
a client to set the CRTC gamma lookup table.

The property is a blob that is interpreted as an array of
drm_color_lut structs of the size specified by the 'GAMMA_LUT_SIZE'
property. Use a linear gamma function and a step function to test
functionality.

BUG=chromium:792757
TEST='atomictest -t crtc_gamma' on eve

Change-Id: I04b481b24aaf2fdecef82f797ec8c0458940f893
Reviewed-on: https://chromium-review.googlesource.com/1030620
Commit-Ready: Daniel Nicoara <dnicoara@chromium.org>
Tested-by: Daniel Nicoara <dnicoara@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>