drm-tests: Add test_out_fence atomic test
This CL adds |out_fence_ptr| property to crtcs and a test
that checks the property works as expected: if set to a pointer,
the pointee will be set to a valid fd, that represents and out fence,
after an atomic commit.

This CL also changes atomictest_property::value to uint64 since
that is the type of drm atomic properties and it is needed to set
the out_fence_ptr property.

BUG=None
TEST=atomictest -t out_fence

Change-Id: I549aefd21d3a19c1ad9a5970e390b939b26170f7
Reviewed-on: https://chromium-review.googlesource.com/1123928
Commit-Ready: Daniele Castagna <dcastagna@chromium.org>
Tested-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>