drm-tests: Add test_in_fence atomic test
This CL adds |IN_FENCE_FD| property to planes and a test that uses it.
The test uses sw_sync, creating a timeline, adding a fence and incrementing
the timeline (that signals the fence).
The fence is set to the primary plane with the atomic commit.

The timeline increment is done on a thread, since calling drmModeAtomicCommit
with an in fence seems to block until the in fence is signalled.

BUG=None
TEST=atomictest -t in_fence

Change-Id: I737bebd846834eeafe6f47a258ad8c2ccb6a3af1
Reviewed-on: https://chromium-review.googlesource.com/1127267
Commit-Ready: Daniele Castagna <dcastagna@chromium.org>
Tested-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>